(refactor) Move api hanler methods from db folder to server foler
