+++ server/api/fetchOngoingProjects.js
@@ -1,5 +1,5 @@
-import model from "./db/models";
-import collection from "./db/collections";
+import model from "../db/models";
+import collection from "../db/collections";
 
 const fetchOngoingProjects = (url, q, res) => {
   const { userId } = q;
